requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
gunicorn>=21.2
psycopg2-binary>=2.9
python-dotenv>=1.0
python-decouple>=3.8
djangorestframework-simplejwt==5.5.1
django-cors-headers>=4.0
dj-database-url>=2.0
django-filter>=23.0
Pillow>=10.0cloudinary>=1.36
django-cloudinary-storage>=0.3.0
whitenoise>=6.6
